requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
Jinja2==3.1.2
itsdangerous==2.0.1
Pillow==10.1.0
numpy
gunicorn==20.1.0
click==8.1.7
MarkupSafe==2.0.1
torch
torchvision